fix circular dependencies, add join button, build errors
