+++ package.json
@@ -4,6 +4,6 @@
 	},
 	"dependencies": {
 		"discord-api-types": "^0.37.17",
-		"discord.js": "^14.7.1"
+		"discord.js": "14.9"
 	}
 }
