+++ source/buttons/startevent.js
@@ -1,6 +1,7 @@
 const { GuildScheduledEventStatus, ButtonBuilder, ButtonStyle, ActionRowBuilder } = require('discord.js');
 const Button = require('../classes/Button.js');
-const { isModerator, getClubDictionary } = require('../helpers.js');
+const { getClubDictionary } = require('../helpers.js');
+const { isModerator } = require("../engines/permissionEngine.js");
 
 const id = "startevent";
 module.exports = new Button(id,
